Allow a resample image merge
Instead of using the "simple" resize algorithm the resample algorithm
can also be used when merging an image into the QR code. When the
resampler is used, we get an anti-aliased merge and therefor a more
pretty final result. The parameter that was added to the merge calls
are always forced to be a boolean when they have to be set. This was
done to be consistent with PHP's handling of "true-ness" for
variables, while still having the internal variable as a strict type.

The workaround for not doing this is generating a large QR code with
a merged image and finaly scaling that back. (Or letting the browser
handle this step)

Another testcase was created to verify that setting the quality value
in the ImageMerge class will actually yield a resampled and different
result from the normal merge.

The calls implemented in the facade have been documented in the
readme and have a small discription in code (where applicable). Some
minor changes were made in the in-code documentation to reflect name
changes etc.